Refactor series-area state handling into SeriesAreaManager.
